# docs/swagger.yaml
definitions:
  controller.DiscoveryPreferencesResponse:
    properties:
      email:
        type: string
      radius_km:
        type: integer
    type: object
...
  controller.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  controller.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  interfaces.DiscoveryPreferencesInput:
    properties:
      radius_km:
        type: integer
    type: object
  interfaces.NotificationSettingsInput:
    properties:
      chat_email:
        type: boolean
      chat_push:
        type: boolean
      expo_push_token:
        type: string
      favorite_email:
        type: boolean
      favorite_push:
        type: boolean
      like_email:
        type: boolean
      like_push:
        type: boolean
      system_email:
        type: boolean
      system_push:
        type: boolean
    type: object
...
  keycloak.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a Keycloak access token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keycloak.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Use refresh token to get a new access token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keycloak.TokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - Authentication
...
  /internal/users/{email}/discovery-preferences:
    get:
      description: Internal endpoint - requires X-Service-Token. Returns discovery
        preferences for a user by email.
      parameters:
      - description: Service token
        in: header
        name: X-Service-Token
        required: true
        type: string
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.DiscoveryPreferencesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get discovery preferences by email (internal)
      tags:
      - DiscoveryPreferences
...
  /internal/users/{email}/notification-settings:
    get:
      description: Internal endpoint for services to get notification settings by
        email
      parameters:
      - description: Service token
        in: header
        name: X-Service-Token
        required: true
        type: string
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get notification settings by email (internal)
      tags:
      - Notifications
...
  /users/me/discovery-preferences:
    get:
      description: Returns the discovery preferences (km radius) for the authenticated
        user. If none exist, defaults are returned.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.DiscoveryPreferencesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get discovery preferences for current user
      tags:
      - DiscoveryPreferences
    put:
      consumes:
      - application/json
      description: Updates the discovery preferences (km radius) for the authenticated
        user.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Discovery preferences input
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.DiscoveryPreferencesInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.DiscoveryPreferencesResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update discovery preferences for current user
      tags:
      - DiscoveryPreferences
...
  /users/me/notification-settings:
    put:
      consumes:
      - application/json
      description: Update notification settings for the authenticated user.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/interfaces.NotificationSettingsInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update notification settings
      tags:
      - Notifications
...
swagger: "2.0"